+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - Introduce `ProductProfuntor` to represent "`SemiProductProfunctor` with a unit"
   - Old `ProductProfunctor#empty` has been deprecated; use `ProductProfunctor#unitP`
   - Old `ProductProfunctor#purePP` has been deprecated; use `ProductProfunctor#pureP`
+* Add `divisedP = diviseP id`
 
 # 0.11.1.1